Refactor in BaseRecordManager to avoid unnecessary code repetition
BaseRecordManager now contains almost all the logic for HeapRecordManager. This is what we called the naive approach that will be used in the cases that the record organization don't provide a better one.

Hash for example can only improve in de (Select/Delete)WhereEquals case
While the orderer can improve WhereEquals and the WhereBetween (if the column is the orderer)